move plugin to ./plugin
